+++ src/main/java/de/marhali/easyi18n/model/bus/ExpandAllListener.java
@@ -1,0 +1,13 @@
+package de.marhali.easyi18n.model.bus;
+
+/**
+ * Single event listener.
+ * @see #onExpandAll()
+ * @author marhali
+ */
+public interface ExpandAllListener {
+    /**
+     * Action to expand all nodes
+     */
+    void onExpandAll();
+}